code-review-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-agent"
version = "0.1.0"
description = "LLM-powered GitHub PR review agent with sandboxed fix verification."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "PyGithub>=2.3.0",
    "pygit2>=1.14.0",
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.21.0",
    "docker>=7.1.0",
    "langgraph>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.setuptools.packages.find]
include = ["agent*", "sandbox*", "evals*"]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
